# openapi.yaml
openapi: 3.0.2
info:
  version: '1.0'
  title: 'LBAW 2Auction Web API'
  description: 'Web Resources Specification for 2Auction'
servers:
- url: http://lbaw.fe.up.pt
  description: Production server
...
paths:
  /login:
    get:
      operationId: R101
      summary: "R101: Login Form"
      description: "Provide Login Form. Access: PUB"
      tags:
        - "M01: Authentication and Individual Profile"
      responses:
        '200':
          description: "Successfully returns the login page view."
          content:
            text/html:
              schema:
                type: string
                example: "<html><head><title>Login</title></head><body>Formulário de Login</body></html>"
    post:
      operationId: R102
      summary: "R102: Login Action."
      description: "Processes the Login form, validating user credentials. If valid, logs the user in. If the user is banned, logs them out and returns an error. Access: PUB"
      tags:
        - "M01: Authentication and Individual Profile"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: "user email"
                password:
                  type: string
                  format: password
                  description: "user password"
              required:
                - email
                - password
      responses:
        '200':
          description: "Authenticated, redirects to intended page."
        '302':
          description: "Redirects to homepage."
        '400':
          description: Validation error on input data.
        '401':
          description: Invalid credentials or banned user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    description: Error message regarding invalid credentials or banned account.
        '500':
          description: Internal server error.
...
  /profile:
    get:
      operationId: R106
      summary: "R106: Display the user profile page"
      description: "Returns a profile page depending on the user's role. Admins will see a page with statistics, and regular users will see their personal profile page. If the user is not authenticated, they are redirected to the home page with a message. Access: OWN"
      tags:
        - 'M01: Authentication and Individual Profile'
      responses:
        '200':
          description: Successfully returns the appropriate profile page view.
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Redirects to the home page if the user is not authenticated, with a "no permissions" message.
          content:
            text/html:
              schema:
                type: string
        '403':
          description: The user is not authorized to access the profile page (if their role is not admin and other conditions are unmet).
        '500':
          description: Internal server error.
...